Using /etc/kvc and /usr/{,local/}/lib/kvc directories
kmods-via-containers starting picking up stuff from /etc/kvc and
/usr/{,local/}/lib/kvc directories. Let's start installing stuff
into those locations. Also parameterize DESTDIR and CONFDIR so
we can control installing into /usr/local/ or into a fakeroot if
we want to. Finally, move all files to the toplevel to simplify
things.

Also dropped spukt as it's not related to mellanox.